Timer counters now 64-bit; Cleaned up floppy code a lot and reverted to single poller; Fixed segment present bit and limit checking at read/write within segment; The ASUS boards now have memregs too; RTC code improved based on suggestion by Sarah Walker; Fixed SVGA odd/even emulation and added chain odd/even support; Removed non-existent CPU's.

+++ src/gameport.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <stdlib.h>
 #include "ibm.h"
 #include "device.h"
@@ -55,7 +56,7 @@ char *joystick_get_button_name(int joystick, int id)
 
 typedef struct gameport_axis_t
 {
-        int count;
+        int64_t count;
         int axis_nr;
         struct gameport_t *gameport;
 } gameport_axis_t;
