+++ src/mouse.h
@@ -1,5 +1,5 @@
 extern void (*mouse_poll)(int x, int y, int b);
 
 extern int mousepos;
-extern int mousedelay;
+extern int64_t mousedelay;
