+++ src/vid_incolor.c
@@ -148,7 +148,7 @@ typedef struct incolor_t
         uint8_t ctrl, ctrl2, stat;
 
         int dispontime, dispofftime;
-        int vidtime;
+        int64_t vidtime;
         
         int firstline, lastline;
